+++ Registration_web/Registration_soft/utils/performance_page.py
@@ -1,11 +1,11 @@
 from Registration_soft.models import PerfData
 from django.http import HttpResponse
 from django.shortcuts import render
-def Performance_manage_in(data):
+def Performance_manage_in(data,uid):
     # from Registration_soft.models import OneTimeDatas
     from Registration_soft.models import Auto_save_Data
     # plname = OneTimeDatas.objects.all().first()
-    date = Auto_save_Data.objects.get(ScrimUniqueId = data['uid'])
+    date = Auto_save_Data.objects.filter(UserUniqueId = uid).get(ScrimUniqueId = data['uid'])
 
     print(data.keys())
     d = data.keys()
@@ -29,18 +29,21 @@ def Performance_manage_in(data):
         PlayerName2 = data['userPlayer2']['pname'],
         PlayerName3 = data['userPlayer3']['pname'],
         PlayerName4 = data['userPlayer4']['pname'],
+        UserUniqueId = uid
+
     )
     Perfomance_data.save()
     
-    Auto_save_Data.objects.filter(ScrimUniqueId = data['uid']).delete()
+    # Auto_save_Data.objects.filter(ScrimUniqueId = data['uid']).delete()
+    Auto_save_Data.objects.filter(UserUniqueId = uid).filter(ScrimUniqueId = data['uid']).delete()
 
 
 
 
-def Performance_manage_show():
+def Performance_manage_show(uid):
     from Registration_soft.models import PerfData
 
-    data = PerfData.objects.all()
+    data = PerfData.objects.filter(ScrimUniqueId = uid).all()
     nod = 0
     show_data = []
     for datas in data :
